Email Templates subnavigation and cleanup.
