+++ app-modules/assistant/src/Filament/Pages/AssistantConfiguration.php
@@ -30,6 +30,7 @@ https://www.canyongbs.com or contact us via email at legal@canyongbs.com.
 
 namespace Assist\Assistant\Filament\Pages;
 
+use App\Filament\Pages\EmailConfiguration;
 use App\Models\User;
 use Filament\Pages\Page;
 use Filament\Navigation\NavigationItem;
@@ -95,7 +96,7 @@ class AssistantConfiguration extends Page
             $subItems = (new AssistantConfiguration())->getSubNavigation();
 
             foreach ($subItems as $subItem) {
-                if (request()->fullUrlIs($subItem->getUrl())) {
+                if (str(request()->fullUrl())->contains(str($subItem->getUrl())->after('/'))) {
                     return true;
                 }
             }
