+++ app-modules/engagement/src/Filament/Resources/EmailTemplateResource.php
@@ -46,6 +46,8 @@ class EmailTemplateResource extends Resource
 
     protected static ?int $navigationSort = 11;
 
+    protected static bool $shouldRegisterNavigation = false;
+
     public static function getPages(): array
     {
         return [
