+++ app-modules/engagement/src/Filament/Resources/EmailTemplateResource/Pages/CreateEmailTemplate.php
@@ -32,7 +32,9 @@ namespace Assist\Engagement\Filament\Resources\EmailTemplateResource\Pages;
 
 use Filament\Forms\Form;
 use FilamentTiptapEditor\TiptapEditor;
+use Filament\Forms\Components\Textarea;
 use Filament\Forms\Components\TextInput;
+use App\Filament\Pages\EmailConfiguration;
 use Filament\Resources\Pages\CreateRecord;
 use Assist\Engagement\Filament\Resources\EmailTemplateResource;
 
@@ -40,18 +42,26 @@ class CreateEmailTemplate extends CreateRecord
 {
     protected static string $resource = EmailTemplateResource::class;
 
+    public function getBreadcrumbs(): array
+    {
+        return [
+            ...(new EmailConfiguration())->getBreadcrumbs(),
+            ...parent::getBreadcrumbs(),
+        ];
+    }
+
     public function form(Form $form): Form
     {
         return $form
+            ->columns(1)
             ->schema([
                 TextInput::make('name')
                     ->string()
-                    ->required(),
-                TextInput::make('description')
-                    ->columnSpanFull()
+                    ->required()
+                    ->autocomplete(false),
+                Textarea::make('description')
                     ->string(),
                 TiptapEditor::make('content')
-                    ->columnSpanFull()
                     ->extraInputAttributes(['style' => 'min-height: 12rem;'])
                     ->required(),
             ]);
