+++ app-modules/engagement/src/Filament/Resources/EmailTemplateResource/Pages/EditEmailTemplate.php
@@ -33,26 +33,36 @@ namespace Assist\Engagement\Filament\Resources\EmailTemplateResource\Pages;
 use Filament\Forms\Form;
 use Filament\Actions\DeleteAction;
 use App\Filament\Fields\TiptapEditor;
+use Filament\Forms\Components\Textarea;
 use Filament\Forms\Components\TextInput;
 use Filament\Resources\Pages\EditRecord;
+use App\Filament\Pages\EmailConfiguration;
 use Assist\Engagement\Filament\Resources\EmailTemplateResource;
 
 class EditEmailTemplate extends EditRecord
 {
     protected static string $resource = EmailTemplateResource::class;
 
+    public function getBreadcrumbs(): array
+    {
+        return [
+            ...(new EmailConfiguration())->getBreadcrumbs(),
+            ...parent::getBreadcrumbs(),
+        ];
+    }
+
     public function form(Form $form): Form
     {
         return $form
+            ->columns(1)
             ->schema([
                 TextInput::make('name')
                     ->string()
-                    ->required(),
-                TextInput::make('description')
-                    ->columnSpanFull()
+                    ->required()
+                    ->autocomplete(false),
+                Textarea::make('description')
                     ->string(),
                 TiptapEditor::make('content')
-                    ->columnSpanFull()
                     ->extraInputAttributes(['style' => 'min-height: 12rem;'])
                     ->required(),
             ]);
