+++ app-modules/engagement/src/Filament/Resources/EmailTemplateResource/Pages/ListEmailTemplates.php
@@ -34,6 +34,7 @@ use Filament\Tables\Table;
 use Filament\Actions\CreateAction;
 use Filament\Tables\Actions\EditAction;
 use Filament\Resources\Pages\ListRecords;
+use App\Filament\Pages\EmailConfiguration;
 use Filament\Tables\Actions\BulkActionGroup;
 use Filament\Tables\Actions\DeleteBulkAction;
 use App\Filament\Columns\OpenSearch\TextColumn;
@@ -42,6 +43,18 @@ use Assist\Engagement\Filament\Resources\EmailTemplateResource;
 class ListEmailTemplates extends ListRecords
 {
     protected static string $resource = EmailTemplateResource::class;
+
+    protected static ?string $navigationIcon = 'heroicon-o-envelope';
+
+    protected static ?string $navigationLabel = 'Email Templates';
+
+    public function getBreadcrumbs(): array
+    {
+        return [
+            ...(new EmailConfiguration())->getBreadcrumbs(),
+            ...parent::getBreadcrumbs(),
+        ];
+    }
 
     public function table(Table $table): Table
     {
@@ -60,6 +73,11 @@ class ListEmailTemplates extends ListRecords
                     DeleteBulkAction::make(),
                 ]),
             ]);
+    }
+
+    public function getSubNavigation(): array
+    {
+        return (new EmailConfiguration())->getSubNavigation();
     }
 
     protected function getHeaderActions(): array
