+++ app-modules/engagement/src/Filament/Resources/SmsTemplateResource/Pages/CreateSmsTemplate.php
@@ -31,8 +31,10 @@ https://www.canyongbs.com or contact us via email at legal@canyongbs.com.
 namespace Assist\Engagement\Filament\Resources\SmsTemplateResource\Pages;
 
 use Filament\Forms\Form;
+use Filament\Forms\Components\Textarea;
 use Filament\Forms\Components\TextInput;
 use Filament\Forms\Components\RichEditor;
+use App\Filament\Pages\EmailConfiguration;
 use Filament\Resources\Pages\CreateRecord;
 use Assist\Engagement\Filament\Resources\SmsTemplateResource;
 
@@ -40,18 +42,26 @@ class CreateSmsTemplate extends CreateRecord
 {
     protected static string $resource = SmsTemplateResource::class;
 
+    public function getBreadcrumbs(): array
+    {
+        return [
+            ...(new EmailConfiguration())->getBreadcrumbs(),
+            ...parent::getBreadcrumbs(),
+        ];
+    }
+
     public function form(Form $form): Form
     {
         return $form
+            ->columns(1)
             ->schema([
                 TextInput::make('name')
                     ->string()
-                    ->required(),
-                TextInput::make('description')
-                    ->columnSpanFull()
+                    ->required()
+                    ->autocomplete(false),
+                Textarea::make('description')
                     ->string(),
                 RichEditor::make('content')
-                    ->columnSpanFull()
                     ->toolbarButtons([
                         'link',
                         'redo',
