+++ app/Filament/Resources/NotificationSettingResource.php
@@ -50,8 +50,7 @@ class NotificationSettingResource extends Resource
 
     public static function getRelations(): array
     {
-        return [
-        ];
+        return [];
     }
 
     public static function getPages(): array
