+++ app/Filament/Resources/NotificationSettingResource/Pages/CreateNotificationSetting.php
@@ -34,6 +34,7 @@ use Filament\Forms\Form;
 use App\Forms\Components\ColorSelect;
 use Filament\Forms\Components\Textarea;
 use Filament\Forms\Components\TextInput;
+use App\Filament\Pages\EmailConfiguration;
 use Filament\Resources\Pages\CreateRecord;
 use App\Filament\Resources\NotificationSettingResource;
 use Filament\Forms\Components\SpatieMediaLibraryFileUpload;
@@ -41,6 +42,14 @@ use Filament\Forms\Components\SpatieMediaLibraryFileUpload;
 class CreateNotificationSetting extends CreateRecord
 {
     protected static string $resource = NotificationSettingResource::class;
+
+    public function getBreadcrumbs(): array
+    {
+        return [
+            ...(new EmailConfiguration())->getBreadcrumbs(),
+            ...parent::getBreadcrumbs(),
+        ];
+    }
 
     public function form(Form $form): Form
     {
